% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_astcat}
\alias{calculate_astcat}
\title{Calculate Aspartate Aminotransferase Category}
\usage{
calculate_astcat(ast, astuln)
}
\arguments{
\item{ast}{Aspartate Aminotransferase (U/L). Numeric vector.}

\item{astuln}{Reference Range Upper Limit (same units as observed values). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Aspartate Aminotransferase Category
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate ASTCAT: 
  if AST [U/L] <= ULN [U/L] then ASTCAT = 0
  if ULN [U/L] < AST [U/L] then ASTCAT = 1
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\%
  mutate(ASTCAT = calculate_astcat(ast = AST, astuln = ASTULN))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(ASTCAT = calculate_astcat())
}
\references{
Ramanathan RK, Egorin MJ, Takimoto CHM, Remick SC, Doroshow JH, LoRusso PA, et al. Phase I and pharmacokinetic study of imatinib mesylate in patients with advanced malignancies and varying degrees of liver dysfunction: a study by the national cancer institute organ dysfunction working group. J Clin Oncol. 2008;26:563-9.

Ramalingam SS, Kummar S, Sarantopoulos J, Shibata S, LoRusso P, Yerk M, et al. Phase I study of vorinostat in patients with advanced solid tumors and hepatic dysfunction: a National Cancer Institute Organ Dysfunction Working Group study. J Clin Oncol. 2010;28(29):4507-12.
}
