% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_bsa}
\alias{calculate_bsa}
\title{Calculate BSA (Body Surface Area)}
\usage{
calculate_bsa(htcm, wtkg)
}
\arguments{
\item{htcm}{Height (cm). Numeric vector.}

\item{wtkg}{Weight (kg). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Body Surface Area (BSA). Body Surface Area is a measure of body fat based on height and weight
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate BSA: 
  BSA = ((HTCM) ^ 0.42246 * (WTKG) ^ 0.51456) * .0235
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(BSA = calculate_bsa(HTCM, WTKG))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(BSA = calculate_bsa())
}
\references{
Gehan EA, George SL. Estimation of human body surface area from height and weight. Cancer Chemother Rep. 1970 Aug;54(4):225-35. PMID: 5527019
}
