% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_crcl}
\alias{calculate_crcl}
\title{Calculate CrCl (Creatinine Clearance)}
\usage{
calculate_crcl(age, wtkg, scr, sexf)
}
\arguments{
\item{age}{Age (y). Numeric vector.}

\item{wtkg}{Weight (kg). Numeric vector.}

\item{scr}{SCR (Serum Creatinine, mg/dL). Numeric vector.}

\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Creatinine Clearance (CrCl), a measure of renal function based on age, weight, gender, and Serum Creatinine. This is the Cockcroft and Gault Calculation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate CRCL: 
  Males: CrCl = ((140 - AGE [year]) × WTKG) ÷ (72 × SCr [mg/dL])
  Females: CrCl = (((140 - AGE [year]) × WTKG) ÷ (72 × SCr[mg/dL])) × 0.85
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(CRCL = calculate_crcl(AGE, WTKG, SCR, SEXF))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(CRCL = calculate_crcl())

# Set a cap at some value, like 160
dmcognigen_cov \%>\% 
  mutate(CRCL = pmin(calculate_crcl(), 160))
}
\references{
Cockcroft DW, Gault MH. Prediction of creatinine clearance from serum creatinine. Nephron. 1976;16:31-41.
}
\seealso{
\code{\link{calculate_crcl_peck}} for Peck formula
}
