% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_egfr}
\alias{calculate_egfr}
\title{Calculate eGFR (Estimated Glomerular Filtration Fate)}
\usage{
calculate_egfr(scr, age, sexf, racen)
}
\arguments{
\item{scr}{SCR (Serum Creatinine, mg/dL). Numeric vector.}

\item{age}{Age (y). Numeric vector.}

\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}

\item{racen}{Race. Numeric vector where the value 2 is expected to correspond to Black/African American.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Estimated Glomerular Filtration Fate (eGFR) for adults
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate EGFR: 
  EGFRMDRD = 175 × (SCR ^ -1.154) × (AGE ^ -0.203) × (0.742 if female) × (1.212 if African American)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(EGFR = calculate_egfr(SCR, AGE, SEXF, RACEN))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(EGFR = calculate_egfr())
}
\references{
\href{www.fda.gov/media/78573/download}{Guidance for Industry: Pharmacokinetics in Patients with Impaired Renal Function — Study Design, Data Analysis, and Impact on Dosing and Labeling}
}
\seealso{
\code{\link{calculate_egfr_child}}
}
