% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_ibw}
\alias{calculate_ibw}
\title{Calculate Ideal Body Weight}
\usage{
calculate_ibw(sexf, htcm)
}
\arguments{
\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}

\item{htcm}{Height (cm). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Ideal Body Weight (IBW), a measure of potential body fat based on height
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate IBW: 
  Males: IBW [kg] = 51.65 [kg] + 1.85 [kg] × ((HTCM [cm] ÷ 2.54) - 60)
  Females: IBW [kg] = 48.67 [kg] + 1.65 [kg] × ((HTCM [cm] ÷ 2.54) - 60)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(IBW = calculate_ibw(sexf = SEXF, htcm = HTCM))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(IBW = calculate_ibw())
}
\references{
Robinson JD, Lupkiewicz SM, Palenik L, Lopez LM, Ariet M. Determination of ideal body weight for drug dosage calculations. AmJHosp Pharm.  1983;40:1016-1019
}
\seealso{
\code{\link{calculate_ibw_child}}
}
