% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_ibw_child}
\alias{calculate_ibw_child}
\title{Calculate Ideal Body Weight for Children}
\usage{
calculate_ibw_child(age, htcm)
}
\arguments{
\item{age}{Age (y). Numeric vector.}

\item{htcm}{Height (cm). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Ideal Body Weight (for children), a measure of potential body fat based on height
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate IBW_CHILD: 
  IBW [kg] = [(HTCM [cm]) ^ 2 ร 1.65] รท 1000
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(IBWCHILD = calculate_ibw_child(age = AGE, htcm = HTCM))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(IBWCHILD = calculate_ibw_child())
}
\references{
Traub SL, Johnson CE. Comparison of methods of estimating creatinine clearance in children. Am J Hosp Pharm. 1980 Feb;37(2):195-201. PMID: 7361791.
}
\seealso{
\code{\link{calculate_ibw}}
}
