% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_lbm}
\alias{calculate_lbm}
\title{Calculate LBM (Lean Body Mass)}
\usage{
calculate_lbm(sexf, wtkg, htcm)
}
\arguments{
\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}

\item{wtkg}{Weight (kg). Numeric vector.}

\item{htcm}{Height (cm). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Lean Body Mass (lbm), an estimation of how much someone weighs without the body fat - how much a persons bones, organs and muscles weigh
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate LBM: 
  Males: LBM = (1.10 × WTKG) - 128 × ((WTKG) ^ 2 ÷ (HTCM) ^ 2)
  Females: LBM = (1.07 × WTKG) - 148 × ((WTKG) ^ 2 ÷ (HTCM) ^ 2)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(LBM = calculate_lbm(SEXF, WTKG, HTCM))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(LBM = calculate_lbm())
}
\references{
Hallynck TH, Soep HH, Thomis JA, et al. Should clearance be normalised to body surface or to lean body mass? Br J Clin Pharmacol. 1981;11:523-526

James WPT. Research on obesity. London. Her Majesty's Stationery Office. 1976
}
