% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_nciliv}
\alias{calculate_nciliv}
\title{Calculate NCI Liver Function Group}
\usage{
calculate_nciliv(tbilcat, astcat)
}
\arguments{
\item{tbilcat}{Total Bilirubin Category. Numeric vector.}

\item{astcat}{Aspartate Aminotransferase Category. Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate NCI Liver Function Group (NCILIV)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate NCILIV: 
  if TBILCAT = 0 and ASTCAT = 0 then NCILIV = 0
  if TBILCAT = 0 and ASTCAT = 1 then NCILIV = 1
  if TBILCAT = 1 then NCILIV = 2
  if TBILCAT = 2 then NCILIV = 3
  if TBILCAT = 3 then NCILIV = 4
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(NCILIV = calculate_nciliv(tbilcat = TBILCAT, astcat = ASTCAT))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(NCILIV = calculate_nciliv())
}
\references{
Ramanathan RK, Egorin MJ, Takimoto CHM, Remick SC, Doroshow JH, LoRusso PA, et al. Phase I and pharmacokinetic study of imatinib mesylate in patients with advanced malignancies and varying degrees of liver dysfunction: a study by the national cancer institute organ dysfunction working group. J Clin Oncol. 2008;26:563-9.

Ramalingam SS, Kummar S, Sarantopoulos J, Shibata S, LoRusso P, Yerk M, et al. Phase I study of vorinostat in patients with advanced solid tumors and hepatic dysfunction: a National Cancer Institute Organ Dysfunction Working Group study. J Clin Oncol. 2010;28(29):4507-12.
}
