% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_rfcat}
\alias{calculate_rfcat}
\title{Calculate Renal Function Category}
\usage{
calculate_rfcat(egfr)
}
\arguments{
\item{egfr}{Estimated Glomerular Filtration Rate (mL/min/1.73m^2). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Renal Function Category (RFCAT)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate RFCAT: 
  1 = Normal Function (>= 90 mL/min)
  2 = Mild Impairment (60-89 mL/min)
  3 = Moderate Impairment (30-59 mL/min)
  4 = Severe Impairment (15-29 mL/min)
  5 = End Stage Disease (< 15 mL/min or Dialysis)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(RFCAT = calculate_rfcat(EGFR))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(RFCAT = calculate_rfcat())
}
\references{
\href{www.fda.gov/media/78573/download}{Guidance for Industry: Pharmacokinetics in Patients with Impaired Renal Function — Study Design, Data Analysis, and Impact on Dosing and Labeling}
}
