% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_tbilcat}
\alias{calculate_tbilcat}
\title{Calculate TBILCAT (TBIL Category)}
\usage{
calculate_tbilcat(tbil, tbiluln)
}
\arguments{
\item{tbil}{Total Bilirubin (mg/dL). Numeric vector.}

\item{tbiluln}{Reference Range Upper Limit (same units as observed values). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate TBIL Category
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate TBILCAT: 
  if TBIL <= ULN then TBILCAT = 0
  if ULN < TBIL <= 1.5 × ULN then TBILCAT = 1
  if 1.5 × ULN < TBIL <= 3 × ULN then TBILCAT = 2
  if TBIL > 3 × ULN then TBILCAT = 3
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(TBILCAT = calculate_tbilcat(TBIL, TBILULN))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(TBILCAT = calculate_tbilcat())
}
\references{
Ramanathan RK, Egorin MJ, Takimoto CHM, Remick SC, Doroshow JH, LoRusso PA, et al. Phase I and pharmacokinetic study of imatinib mesylate in patients with advanced malignancies and varying degrees of liver dysfunction: a study by the national cancer institute organ dysfunction working group. J Clin Oncol. 2008;26:563-9.

Ramalingam SS, Kummar S, Sarantopoulos J, Shibata S, LoRusso P, Yerk M, et al. Phase I study of vorinostat in patients with advanced solid tumors and hepatic dysfunction: a National Cancer Institute Organ Dysfunction Working Group study. J Clin Oncol. 2010;28(29):4507-12.
}
