% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt.R
\name{cnt}
\alias{cnt}
\title{Count observations and unique values by group}
\usage{
cnt(
  .data,
  ...,
  n_distinct_vars = NULL,
  n_distinct_combined = TRUE,
  n_cumulative = TRUE,
  prop = FALSE,
  pct = FALSE
)
}
\arguments{
\item{.data}{data frame}

\item{...}{variables to group by}

\item{n_distinct_vars}{variables to count the unique values of}

\item{n_distinct_combined}{logical indicating whether to count the unique
values of the combinations of \code{n_distinct_vars}}

\item{n_cumulative}{logical indicating whether to include a cumulative sum
variable named "n_cumulative"}

\item{prop}{logical indicating whether to include a proportion variable named
"prop"}

\item{pct}{logical indicating whether to include a percentage variable named
"pct"}
}
\value{
data frame
}
\description{
This function is an extension of \code{dplyr::count}. Counts are produced for
unique combinations of grouping variables; optionally includes proportions,
percentages, and cumulative counts. If any \code{n_distinct_vars} are
provided, counts of the unique values are produced for each.
}
\examples{
# count unique values of `am` along with the number of distinct values of
# `carb`, `cyl`, and their combination.
mtcars \%>\% 
  cnt(am, n_distinct_vars = c(carb, cyl))
}
