% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_environment_data.R
\name{cnt_search_result}
\alias{cnt_search_result}
\title{Count search results}
\usage{
cnt_search_result(
  search_result,
  n_distinct_vars = NULL,
  n_distinct_combined = TRUE,
  df_names = NULL,
  extra_vars = NULL,
  ignore_vars = NULL
)
}
\arguments{
\item{search_result}{a \code{search_result} object obtained from
[search_environment_data()].}

\item{n_distinct_vars}{variables to count the unique values of}

\item{n_distinct_combined}{logical indicating whether to count the unique
values of the combinations of \code{n_distinct_vars}}

\item{df_names}{an optional character vector of data.frame names to subset
to.}

\item{extra_vars, ignore_vars}{optional character vectors of variables to
include/exclude in [cnt()] results.}
}
\value{
a named list of [cnt()] results.
}
\description{
When patterns are matched in variable names or labels, no subset is
performed. When patterns are only matched in variable content, the
data.frames are subset to rows where some variable matches the pattern.
}
\examples{
data(dmcognigen_cov)
study_search <- search_environment_data(
  stringr::regex("study", ignore_case = TRUE)
)
cnt_search_result(
  study_search, 
  n_distinct_vars = "USUBJID", 
  extra_vars = "ACTARM"
)
}
\seealso{
[search_environment_data()]
}
