% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_tbls.R
\name{decode_tbls}
\alias{decode_tbls}
\alias{as_decode_tbl}
\alias{as_decode_tbls}
\title{Build decode tables}
\usage{
as_decode_tbl(x = tibble::tibble())

as_decode_tbls(x = list())
}
\arguments{
\item{x}{A tibble or data frame for \code{as_decode_tbl()}. A list for
\code{as_decode_tbls()}.}
}
\value{
a \code{decode_tbl} or \code{decode_tbls} object.
}
\description{
Objects of class \code{decode_tbl} include all pairs of
  \code{lvl} and \code{lbl} for a single \code{var}. They are generated based
  on variable definitions or values within a dataset. 

Objects of \code{decode_tbls} are lists of \code{decode_tbl} objects.
}
\seealso{
\code{\link{print.decode_tbl}}, \code{\link{extract_decode_tbls}},
\code{\link{extract_decode_tbls_from_data}}, \code{\link{join_decodes}}
}
