% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{describe_param}
\alias{describe_param}
\title{pull details from params table}
\usage{
describe_param(.param)
}
\arguments{
\item{.param}{name of the parameter to describe}
}
\description{
This function is mostly internal, but is used for documentation and is
therefore exported.
}
\keyword{internal}
