% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmcognigen_conc}
\alias{dmcognigen_conc}
\title{PK Concentrations Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3556 rows and 29 columns.
}
\usage{
dmcognigen_conc
}
\description{
A dataset generated from \code{pharmaversesdtm} example datasets, which are
updated SDTM datasets that use data from the CDISC pilot project.
}
\seealso{
\code{\link[pharmaversesdtm]{pc}}
}
\keyword{datasets}
