% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmcognigen_cov}
\alias{dmcognigen_cov}
\title{Covariate Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 53 columns.
}
\usage{
dmcognigen_cov
}
\description{
A dataset generated from \code{pharmaversesdtm} example datasets, which are
updated SDTM datasets that use data from the CDISC pilot project.
}
\seealso{
\code{\link[pharmaversesdtm]{dm}}, 
\code{\link[pharmaversesdtm]{lb}}, 
\code{\link[pharmaversesdtm]{vs}}
}
\keyword{datasets}
