% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmcognigen_pk}
\alias{dmcognigen_pk}
\title{PK Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3556 rows and 105 columns.
}
\usage{
dmcognigen_pk
}
\description{
A dataset generated from \code{pharmaversesdtm} example datasets, which are
updated SDTM datasets that use data from the CDISC pilot project.
}
\seealso{
\code{\link[pharmaversesdtm]{dm}}, 
\code{\link[pharmaversesdtm]{ex}},
\code{\link[pharmaversesdtm]{lb}}, 
\code{\link[pharmaversesdtm]{pc}},
\code{\link[pharmaversesdtm]{vs}}
}
\keyword{datasets}
