% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmcognigen_pk_requirements}
\alias{dmcognigen_pk_requirements}
\title{Example PK Data Requirements}
\format{
An object of class \code{requirements} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 54 rows and 5 columns.
}
\usage{
dmcognigen_pk_requirements
}
\description{
Data Requirements designed based on \code{pharmaversesdtm} example datasets,
which are updated SDTM datasets that use data from the CDISC pilot project.

This has the same structure and attributes as results from
\code{\link{read_requirements}} and \code{\link{as_requirements}}.
}
\examples{
head(dmcognigen_pk_requirements)

as_decode_tbls(dmcognigen_pk_requirements)

}
\keyword{datasets}
