% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_decode_tbls.R
\name{extract_decode_tbls}
\alias{extract_decode_tbls}
\alias{extract_decode_tbl}
\title{Extract decode tables}
\usage{
extract_decode_tbl(variable_name, decode)

extract_decode_tbls(variable_name, decode)
}
\arguments{
\item{variable_name}{A \code{character} vector of variable names.}

\item{decode}{A \code{character} vector. Items are expected to be separated
by new lines. Name-value pairs should be separated by a single equal sign.}
}
\description{
This family of functions can be used to extract
  \code{\link{decode_tbls}} from vectors of variable names and decodes.
  Generally, users are expected to use \code{extract_decode_tbls}.

\describe{
  \item{\code{extract_decode_tbl}}{Generates a \code{decode_tbl} for one
  variable name and set of decodes.}
  \item{\code{extract_decode_tbls}}{Generates a \code{decode_tbls} list given
  vectors of variable names and decodes. Results with no identified decodes
  are dropped.}
}
}
\examples{
possible_decodes <- tibble::tribble(
  ~var,     ~decode,
  # compact
  "SEXF",   "0=Male\n1=Female",                                                        
  # yes/no without single quotes
  "LGL",    "Y=Yes\nN=No",                                                             
  # yes/no with single quotes
  "LGLQ",   "'Y'=Yes\n'N'=No",                                                         
  # with spaces and numbers
  "AGEPED", "0=Neonate to 1 month\n1=Infant: 1 month to 2 years",                      
  # with parentheses in decode
  "MDV",    "0=PK or PD measure\n1=Dose or Other(EVID=2)",                             
  # with special characters and parentheses
  "RFCAT",  "1=Normal (>=90 mL/min)\n2=Mild (60-89 mL/min)\n3=Moderate (30-59 mL/min)",
  # item that does not include encoding
  "BAD",    "This is not a decode"                                                     
)

extract_decode_tbls(possible_decodes$var, possible_decodes$decode)

possible_decodes |> 
  with(extract_decode_tbls(var, decode))

}
\seealso{
\code{\link{extract_decode_tbls_from_data}} to extract
\code{\link{decode_tbls}} from a dataset;
\code{\link{join_decode_labels}} and \code{\link{join_decode_levels}} to
merge \code{\link{decode_tbls}} to a dataset.
}
