% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_decode_tbls_from_data.R
\name{extract_decode_tbls_from_data}
\alias{extract_decode_tbls_from_data}
\title{Extract decode tables from data}
\usage{
extract_decode_tbls_from_data(.data, lvl_to_lbl = "{var}C")
}
\arguments{
\item{.data}{data frame}

\item{lvl_to_lbl}{a named \code{list}, where elements correspond to matching
names in \code{.data} and \code{decode_tbls}. One element can be left
unnamed to provide the default transformation. Each element should be
either: (1) a function/formula; (2) a custom glue specification. The
default is \code{"{var}C"}, which means the label variable name is the
level variable name appended with a "C".}
}
\value{
\code{\link{decode_tbls}}
}
\description{
Extract decode tables from data
}
\examples{
dmcognigen_cov |> 
  extract_decode_tbls_from_data()

dmcognigen_cov |> 
  extract_decode_tbls_from_data(
    lvl_to_lbl = c(
    # specify mapping from lvl to lbl by name
    "RACEN" = "RACE", 
    # one unnamed, default mapping can be provided
    ~ stringr::str_remove(.x, "CD$"))
  )
}
