% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_decodes.R
\name{extract_decodes}
\alias{extract_decodes}
\alias{extract_decode_tbl}
\alias{extract_decode_tbls}
\title{Interact with decodes and their coded values}
\usage{
extract_decode_tbl(variable_name, decode)

extract_decode_tbls(variable_name, decode)
}
\arguments{
\item{variable_name}{A \code{character} vector of variable names.}

\item{decode}{A \code{character} vector. Items are expected to be separated
by new lines. Name-value pairs should be separated by a single equal sign.}
}
\description{
This family of functions allows the user to transform name-value
pairs of coded values in a variety of ways.

\describe{
  \item{\code{extract_decode_tbl}}{Generates a \code{decode_tbl} for one
  variable name and set of decodes.}
  \item{\code{extract_decode_tbls}}{Generates a \code{decode_tbls} list given
  vectors of variable names and decodes. Results with no identified decodes
  are dropped.}
}
}
\examples{
possible_decodes <- tibble::tribble(
  ~var,     ~decode,
  # compact
  "SEXF",   "0=Male\n1=Female",                                                        
  # yes/no without single quotes
  "LGL",    "Y=Yes\nN=No",                                                             
  # yes/no with single quotes
  "LGLQ",   "'Y'=Yes\n'N'=No",                                                         
  # with spaces and numbers
  "AGEPED", "0=Neonate to 1 month\n1=Infant: 1 month to 2 years",                      
  # with special characters (<)
  "AGECAT", "0=<65 years\n1=65 to 74 years\n2=75 years or older",                      
  # with parentheses in decode
  "MDV",    "0=PK or PD measure\n1=Dose or Other(EVID=2)",                             
  # with special characters and parentheses
  "RFCAT",  "1=Normal (>=90 mL/min)\n2=Mild (60-89 mL/min)\n3=Moderate (30-59 mL/min)",
  # item that does not include encoding
  "BAD",    "This is not a decode"                                                     
)

extract_decode_tbls(possible_decodes$var, possible_decodes$decode)
}
