% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_from_calculations_table}
\alias{get_from_calculations_table}
\title{pull details from calculations_table based on provided arguments}
\usage{
get_from_calculations_table(.var_name, .field)
}
\arguments{
\item{.var_name}{variable name to extract a field from}

\item{.field}{field to extract}
}
\description{
This function is mostly internal, but is used for documentation and is
therefore exported.
}
\keyword{internal}
