% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_which.R
\name{in_which}
\alias{in_which}
\title{Find objects containing names}
\usage{
in_which(x, envir = .GlobalEnv, ignore_case = TRUE)
}
\arguments{
\item{x}{a \code{character} vector where matches to any elements are sought.}

\item{envir}{which environment to use in listing the available objects.
Defaults to the global environment. This can also be a \code{list}
containing objects to check.}

\item{ignore_case}{\code{logical} representing whether to ignore case.}
}
\value{
returns a named \code{list} of \code{character} vectors containing
  the names of \code{list} objects with some names matching to \code{x} or
  invisibly returns \code{NULL} where there are no \code{list} objects in
  \code{envir}.
}
\description{
Find \code{list} objects whose names contain any elements of \code{x}. Note
that \code{data.frame} objects are considered \code{list} objects.
}
\examples{
data(mtcars); in_which("cyl")

# check for common subject identifiers, explicitly defining case
in_which(
  c("ID", "SUBJECT", "SUBJID", "USUBJID"), 
  ignore_case = FALSE
)
}
