% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_tbl.R
\name{print.decode_tbl}
\alias{print.decode_tbl}
\alias{print.decode_tbls}
\title{Print values of decode tables}
\usage{
\method{print}{decode_tbl}(x, align = c("left", "center", "right"), ...)

\method{print}{decode_tbls}(x, align = c("left", "center", "right"), ...)
}
\arguments{
\item{x}{A \code{decode_tbl} or \code{decode_tbls} object.}

\item{align}{A string for the alignment of the output. Either "left"
(default), "center" (aligns '='), or "right".}

\item{...}{Passed to \code{print}.}
}
\description{
Print values of decode tables
}
