% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_environment_data.R
\name{print.search_result}
\alias{print.search_result}
\title{Print search results}
\usage{
\method{print}{search_result}(x, ...)
}
\arguments{
\item{x}{a \code{search_result} object}

\item{...}{Passed to \code{print}.}
}
\description{
Print search results
}
