% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx_tables.R
\name{read_docx_tables}
\alias{read_docx_tables}
\title{Read tables from a Word file (docx)}
\usage{
read_docx_tables(path, docx_header_pattern = NULL, combine = TRUE, ...)
}
\arguments{
\item{path}{the path to a Word file (docx).}

\item{docx_header_pattern}{one or more patterns of required table header
names. Can be \code{character} or a list containing any combination of
\code{character}, \code{\link[stringr]{regex}}, and
\code{\link[stringr]{fixed}} patterns. For case-insensitive, use regex.}

\item{combine}{\code{logical} indicating whether to combine the tables into
one data frame or return a list of tables.}

\item{...}{optional arguments passed to \code{\link[docxtractr]{read_docx}}.}
}
\value{
a data frame when \code{combined = TRUE}; a list of data frames when
  \code{combined = FALSE}.
}
\description{
Read tables from a Word file (docx)
}
\examples{
\dontrun{

# read and combine all tables from a docx.
read_docx_tables("results.docx")

# read and combine tables from a docx that match a pattern.
read_docx_tables(
  path = "requirements.docx",
  docx_header_pattern = list(
    stringr::regex("variable", ignore_case = TRUE)
  )
)

# optionally, read in with tracked changes accepted or rejected.
read_docx_tables(
  path = "requirements.docx",
  track_changes = "accept"
)
}

}
\seealso{
\code{\link[docxtractr]{read_docx}}, 
\code{\link[docxtractr]{docx_extract_all_tbls}}
}
