% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_environment_data.R
\name{search_environment_data}
\alias{search_environment_data}
\title{Search data in an environment for a pattern}
\usage{
search_environment_data(pattern, envir = .GlobalEnv)
}
\arguments{
\item{pattern}{A pattern to search for. Can be \code{\link[stringr]{regex}},
\code{\link[stringr]{fixed}}, or a single character pattern.}

\item{envir}{The environment or named list to search within.}
}
\value{
a \code{search_result} object.
}
\description{
Searches variable names, variable labels, and variable content.
}
\examples{
data(mtcars)
search_environment_data("cyl")

data(dmcognigen_cov)
study_search <- search_environment_data(
  stringr::regex("study", ignore_case = TRUE)
)
cnt_search_result(study_search)
}
\seealso{
[cnt_search_result()]
}
