% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_decode_factors.R
\name{set_decode_factors}
\alias{set_decode_factors}
\title{Create factor variables based on decodes}
\usage{
set_decode_factors(.data, decode_tbls, new_names = NULL, ...)
}
\arguments{
\item{.data}{data frame}

\item{decode_tbls}{\code{decode_tbls} list as returned by
\code{extract_decode_tbls()} or \code{extract_decode_tbls_from_data()}.
Alternatively, the result of \code{read_requirements()} can be used
directly.}

\item{new_names}{\code{NULL} to update in-place or a named \code{list}, where
elements correspond to matching names in \code{.data} and
\code{decode_tbls}. One element can be left unnamed to provide the default
transformation. Each element should be either: (1) a function/formula; (2)
a custom glue specification, like \code{"{var}CAT"} to create a new
variable that is the original variable name appended with "CAT" for
categorical. The default is \code{NULL}, to update in-place.}

\item{...}{additional arguments (currently unused)}
}
\value{
data frame with the same number of rows as \code{.data}. New/modified
  variables will be included for matching content in \code{.data} and
  \code{decode_tbls}.
}
\description{
Create factor variables based on decodes
}
\examples{

# create factors in-place
dmcognigen_cov |> 
  set_decode_factors(dmcognigen_pk_requirements) |>
  cnt(SEXF, RACEN)

# or use a mapping convention to define new names
dmcognigen_cov |> 
  cnt(SEXF, RACEN) |>
  set_decode_factors(dmcognigen_pk_requirements, new_names = c(SEXF = "SEXFCAT"))
  
dmcognigen_cov |> 
  cnt(SEXF, RACEN) |>
  set_decode_factors(dmcognigen_pk_requirements, new_names = "{var}CAT")

}
