% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_labels}
\alias{set_labels}
\alias{set_labels<-}
\title{Set variable labels}
\usage{
set_labels(data, labels = list(), ...)

set_labels(data, ...) <- value
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble}}

\item{labels, value}{either a named \code{list} or \code{character} vector, a
\code{requirements} object, or a data frame with labelled variables.
Additionally, \code{NULL} can be used to remove variable labels.}

\item{...}{named pairs where the name corresponds to one variable in
\code{data} and the value is either the variable label or a source to 
inherit the label from. This label is kept over an element in \code{labels}.}
}
\value{
object with the same structure as \code{data} with label attributes
  assigned to the variables provided in \code{labels} and \code{...}.
}
\description{
Set variable labels
}
\examples{

# remove all variable labels for examples
pk_no_labels <- dmcognigen_pk |>
  set_labels(NULL)

# with a requirements object
pk_with_labels <- pk_no_labels |>
  set_labels(dmcognigen_pk_requirements)

# with another data set
pk_with_labels <- pk_no_labels |>
  set_labels(dmcognigen_pk)
  
# by providing individual variable labels with `...`
pk_with_labels <- pk_no_labels |>
  set_labels(ID = "ID")

# or any combination
pk_with_labels <- pk_no_labels |>
  set_labels(dmcognigen_pk_requirements, USUBJID = dmcognigen_cov, ID = "ID")

# apply labels from multiple datasets by using the function multiple times
pk_with_labels <- pk_no_labels |>
  set_labels(dmcognigen_cov) |>
  set_labels(dmcognigen_dose) |>
  set_labels(dmcognigen_conc)
}
