% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_variables.R
\name{stationary_variables}
\alias{stationary_variables}
\alias{distinct_stationary_variables}
\title{Subset to stationary content (one row per group)}
\usage{
distinct_stationary_variables(.data, ...)

stationary_variables(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{grouping variables to use to determine whether other variables are
stationary. If omitted, will subset to variables that are stationary
throughout the entire data frame. Any existing groups are ignored.}
}
\value{
An object of the same type as \code{.data} with fewer or equal
  columns, fewer or equal rows, and no groups.
}
\description{
\code{distinct_stationary_variables()} subsets data to one row per group
containing only the variables that are constant in all groups.
\code{stationary_variables()} returns the names of these variables.

The main difference compared to \code{\link[dplyr]{distinct}} is that this
function keeps all stationary variables, while \code{\link[dplyr]{distinct}}
keeps either the variables in \code{...} or all variables.
}
\examples{
library(dplyr)

# already one row per group, no change.
identical(
  dmcognigen_cov,
  dmcognigen_cov |>
    distinct_stationary_variables(USUBJID)
)

# all constant variables (no groups)
dmcognigen_cov |>
  distinct_stationary_variables()

# names of all constant variable within groups
dmcognigen_cov |>
  stationary_variables(RACEN, SEXF)

# magrittr pipe lets us pass data as `.`
# so we can modify data in a pipeline and reference the resulting stationary variables.
dmcognigen_dose \%>\%
  select(1:10) \%>\%
  cnt(across(stationary_variables(.)), n_distinct_vars = USUBJID)

# or can reference the stationary variables in another data frame.
# 
# count all stationary dose-related variables
dmcognigen_pk \%>\%
  select(STUDYID, USUBJID, any_of(stationary_variables(dmcognigen_dose))) \%>\%
  cnt(across(stationary_variables(., DVID)), n_distinct_vars = USUBJID)

# count all stationary concentration-related variables
dmcognigen_pk \%>\%
  select(STUDYID, USUBJID, any_of(stationary_variables(dmcognigen_conc))) \%>\%
  cnt(across(stationary_variables(., DVID)), n_distinct_vars = USUBJID)
}
