% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate Body Mass Index}
\usage{
calculate_bmi(htcm, wtkg)
}
\arguments{
\item{htcm}{Height (cm). Numeric vector.}

\item{wtkg}{Weight (kg). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Body Mass Index (BMI). BMI is a measure of body fat based on height and weight
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate BMI: 
  BMI [kg/m^2] = WTKG [kg] ÷ (HTCM [cm] ÷ 100) ^ 2
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(BMI = calculate_bmi(htcm = HTCM, wtkg = WTKG))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(BMI = calculate_bmi())
}
\references{
\href{https://www.cdc.gov/healthyweight/assessing/bmi/adult_bmi/index.html}{CDC}
}
