% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_crcl_peck}
\alias{calculate_crcl_peck}
\title{Calculate CrCl peck (Creatinine Clearance using the Peck Formula)}
\usage{
calculate_crcl_peck(age, wtkg, crcl, ibw, scr, sexf)
}
\arguments{
\item{age}{Age (y). Numeric vector.}

\item{wtkg}{Weight (kg). Numeric vector.}

\item{crcl}{Creatinine Clearance}

\item{ibw}{Ideal Body Weight (kg). Numeric vector.}

\item{scr}{SCR (Serum Creatinine, mg/dL). Numeric vector.}

\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Creatinine Clearance (CrCl) using the Peck Formula, a measure of renal function based on age, ideal weight, gender, and Serum Creatinine
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate CRCL_PECK: 
  If WTKG < 120\% of IBW, CrCl = CRCL (Cockroft & Gault)
  If WTKG >= 120\% of IBW:
    Males: CrCl = ((140 - AGE) × IBW) ÷ (72 × SCr)
    Females: CrCl = (((140 - AGE) × IBW) ÷ (72 × SCr)) × 0.85
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(CRCLP = calculate_crcl_peck(AGE, WTKG, CRCL, IBW, SCR, SEXF))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(CRCLP = calculate_crcl_peck())
}
\references{
Peck CC, Conner DP, Murphy MG. Simple Techniques for Individualizing Drug Therapy. Vancouver, WA. Applied Therapeutics, Inc. 1989.
}
\seealso{
\code{\link{calculate_crcl}} for Cockcroft and Gault formula
}
