% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_egfr}
\alias{calculate_egfr}
\title{Calculate Estimated Glomerular Filtration Rate (eGFR)}
\usage{
calculate_egfr(age, racen, scr, sexf)
}
\arguments{
\item{age}{Age (y). Numeric vector.}

\item{racen}{Race. Numeric vector where the value 2 is expected to correspond to Black/African American.}

\item{scr}{Serum Creatinine (mg/dL). Numeric vector.}

\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Estimated Glomerular Filtration Rate (eGFR) for adults
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate EGFR: 
  EGFR [mL/min/1.73m^2] = 175 × (SCR [mg/dL] ^ -1.154) × (AGE [y] ^ -0.203) × (0.742 if female) × (1.212 if African American)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(EGFR = calculate_egfr(
    age = AGE, 
    racen = RACEN,
    scr = SCR, 
    sexf = SEXF
  ))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(EGFR = calculate_egfr())
}
\references{
\href{https://www.fda.gov/media/78573/download}{Guidance for Industry: Pharmacokinetics in Patients with Impaired Renal Function — Study Design, Data Analysis, and Impact on Dosing and Labeling}
}
\seealso{
\code{\link{calculate_egfr_child}}
}
