% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_egfr_child}
\alias{calculate_egfr_child}
\title{Calculate eGFR (Estimated Glomerular Filtration Fate for children) for
children using the Schwartz formula}
\usage{
calculate_egfr_child(htcm, scr, age, sexf)
}
\arguments{
\item{htcm}{Height (cm). Numeric vector.}

\item{scr}{SCR (Serum Creatinine, mg/dL). Numeric vector.}

\item{age}{Age (y). Numeric vector.}

\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Estimated Glomerular Filtration Fate (eGFR) for Pedatrics using the Schwartz formula
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate EGFR_CHILD: 
  Children (AGE <= 16), where:
  egfr_child = (k * htcm) / scr (mg/dL)
  k = 0.45 for AGE <= 1 (full term)
  k = 0.55 for girls 1 < AGE <= 16
  k = 0.55 for boys 1 < AGE < 13
  k = 0.70 for boys 13 <= AGE <= 16
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(EGFRSCHW = calculate_egfr_child(HTCM, SCR, AGE, SEXF))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(EGFRSCHW = calculate_egfr_child())
}
\references{
Schwartz GJ, Furth SL. Glomerular filtration rate measurement and estimation in chronic kidney disease. Pediatr Nephrol. 2007;22(11):1839-48.
}
\seealso{
\code{\link{calculate_egfr}}
}
