% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_lbm}
\alias{calculate_lbm}
\title{Calculate Lean Body Mass}
\usage{
calculate_lbm(htcm, sexf, wtkg)
}
\arguments{
\item{htcm}{Height (cm). Numeric vector.}

\item{sexf}{Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.}

\item{wtkg}{Weight (kg). Numeric vector.}
}
\value{
This function returns a numeric vector the same length as its inputs
}
\description{
Calculate Lean Body Mass (LBM), an estimation of how much someone weighs without the body fat; how much a persons bones, organs and muscles weigh
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Formula to calculate LBM: 
  Males: LBM [kg] = (1.10 × WTKG [kg]) - 128 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
  Females: LBM [kg] = (1.07 × WTKG [kg]) - 148 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
}\if{html}{\out{</div>}}
}
\examples{
library(dplyr)

dmcognigen_cov \%>\% 
  mutate(LBM = calculate_lbm(
    htcm = HTCM,
    sexf = SEXF, 
    wtkg = WTKG 
  ))

# Below will also work if the dataset contains expected variables
dmcognigen_cov \%>\% 
  mutate(LBM = calculate_lbm())
}
\references{
Hallynck TH, Soep HH, Thomis JA, et al. Should clearance be normalised to body surface or to lean body mass? Br J Clin Pharmacol. 1981;11:523-526

James WPT. Research on obesity. London. Her Majesty's Stationery Office. 1976
}
