% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_labels.R
\name{set_labels}
\alias{set_labels}
\alias{set_labels<-}
\title{Set variable labels}
\usage{
set_labels(data, labels)

set_labels(data) <- value
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble}}

\item{labels}{named \code{list} or \code{character} vector. Alternatively, 
the result of \code{read_requirements()} can be used directly.}

\item{value}{(labels) named \code{list} or \code{character} vector}
}
\value{
object with the same structure as \code{data}, with label attributes
  assigned to variables provided in \code{labels}.
}
\description{
Set variable labels
}
\examples{
str(mtcars)

custom_labels <- list(
  mpg = "Miles per gallon", 
  cyl = "Number of cylinders"
)

# set labels for an existing object in-place
custom_mtcars <- mtcars
set_labels(custom_mtcars) <- custom_labels
str(custom_mtcars)

# or assign a labelled dataset to an object
labelled_mtcars <- set_labels(mtcars, custom_labels)
str(labelled_mtcars)

# or with the pipe
labelled_mtcars <- mtcars \%>\% 
  set_labels(custom_labels)
str(labelled_mtcars)
}
